requirements.txt
djangorestframework
djangorestframework-jwt
django-cors-headers